Fixed MWMHints struct in 64-bit systems. (Thanks to Roland Scheidegger)
there's a problem with the MWMHints definition in vdr-xineliboutput. It
does not work on x86_64 at least with metacity and kwin, as these use a
different definition (hence once borders are toggled off they are gone
for good).
This is a strange issue there seems to be no reference source what the
definition of this struct should be, some quick search finds 3
incompatible versions (on 64bit archs) and a dozen or so more compatible
ones (only differing in signedness).
In any case, metacity, kwin agree, and this is also what mplayer uses -
xine though itself also uses the same version as xineliboutput which
doesn't work with at least metacity/kwin. It's quite likely some obscure
wm out there expects one of the other definitions (I think fvwm is same
as metacity, while enlightenment and afterstep need the all-32bit
definition). What a mess.
Anyway this patch just changes the definition to what's more likely to work.

+++ xine_sxfe_frontend.c
@@ -4,7 +4,7 @@
  * See the main source file 'xineliboutput.c' for copyright information and
  * how to reach the author.
  *
- * $Id: xine_sxfe_frontend.c,v 1.199 2012-02-08 11:37:08 phintuka Exp $
+ * $Id: xine_sxfe_frontend.c,v 1.200 2012-02-17 07:44:25 phintuka Exp $
  *
  */
 
@@ -72,11 +72,11 @@
 #define MWM_HINTS_DECORATIONS       (1L << 1)
 #define PROP_MWM_HINTS_ELEMENTS     5
 typedef struct _mwmhints {
-  uint32_t     flags;
-  uint32_t     functions;
-  uint32_t     decorations;
-  int32_t      input_mode;
-  uint32_t     status;
+  unsigned long flags;
+  unsigned long functions;
+  unsigned long decorations;
+  long          input_mode;
+  unsigned long status;
 } MWMHints;
 
 #ifdef HAVE_XRENDER
